fix: restore the reference bf16 dtype flow (rope island, t_emb cast, add_noise)
The reference runs weight_dtype = torch.bfloat16 end-to-end. The MLX port
dropped the two deliberate fp32-island cast-backs and let fp32 tables
contaminate the whole forward via MLX's strict type promotion (smeltr
session void-dogfood-bf16: 182 s of float32 gemms + 71k bf16->fp32
weight upcasts on a 30-step VOID run):

- embeddings.apply_rotary_emb: compute the rotation in fp32 and cast the
  result back to x.dtype (mirror of diffusers embeddings.py:1231).
- cogvideox_transformer3d: t_emb cast to hidden_states.dtype after
  time_proj (mirror of the reference :597) — the fp32 temb contaminated
  every block's modulation path.
- scheduler.add_noise: alphas cast to the sample dtype before the sqrt
  (mirror of diffusers add_noise). step() deliberately KEEPS the torch
  fp32-promotion behavior — the reference pipeline casts latents back
  after each step (pipeline_cogvideox_fun_inpaint.py:1170), and a test
  pins that contract so nobody "fixes" the scheduler blindly.

Also: anchor the `models*` gitignore pattern to the repo root (it blocked
git add on the videox_fun_mlx/models source dir) and stop tracking
__pycache__ artifacts.

5 new dtype-flow tests (rope island semantics, scheduler contracts, tiny
transformer bf16 end-to-end); full suite 53 green. Validated end-to-end
through the VOID two-pass run: fp32 gemms 190 s -> 0, wall 689 -> 532 s,
PSNR fp32-vs-bf16 output 38.5 dB.

+++ videox_fun_mlx/models/embeddings.py
@@ -278,4 +278,10 @@ def apply_rotary_emb(
     else:
         raise ValueError(f"`use_real_unbind_dim={use_real_unbind_dim}` but should be -1 or -2.")
 
-    return x * cos + x_rotated * sin
+    # Mirror diffusers: compute the rotation in float32, cast the result
+    # back to x.dtype (embeddings.py:1231 — the deliberate fp32 island).
+    # Without the cast, fp32 cos/sin tables promote bf16 q/k to fp32 and
+    # contaminate the whole attention (MLX strict promotion).
+    return (
+        x.astype(mx.float32) * cos + x_rotated.astype(mx.float32) * sin
+    ).astype(x.dtype)
